# _addOperation calls are serialized