https://github.com/orbitdb/orbit-db/issues/737

adds a queue to _addOperation()
close() awaits for the queue to empty before resolving