@tabcat 2020-03-27
pretty sure i made a serious mistake in the original commit that would have made the queue per all instances instead of queue per store instance.