@vaultec81 2020-03-27
I suggest making a simple test that runs the code as if it would be a end user, or checking to see if the results were as intended this._addOperations calls are processed sequentially.