@haadcode 2020-03-31
The queue will be on critical code path. Before adding it, merging the PR, we should understand how this effects performance.

Can you run the benchmarks and provide the results for comparison?