@tabcat 2020-03-31
#### no queue:

60s benchmark-add: 
```
--> Average of 520.9 q/s in the last 10 seconds
511 queries per second, 31466 queries in 60 seconds (Oplog: 31466)
```
benchmark-load:
```
Starting IPFS...
DB entries: 1000
Writing DB...
writing took 2715 ms
Loading DB...
load took 1369 ms
```

#### with queue:

60s benchmark-add:
```
--> Average of 551.8 q/s in the last 10 seconds
554 queries per second, 32507 queries in 60 seconds (Oplog: 32507)
```
benchmark-load:
```
Starting IPFS...
DB entries: 1000
Writing DB...
writing took 2462 ms
Loading DB...
load took 1260 ms
```

they were pretty back and forth