@tabcat 2020-03-31
#### without queue:
benchmark-add ran 60s 5 times:
```
1: 33193 queries in 60 seconds
2: 32637 queries in 60 seconds
3: 32104 queries in 60 seconds
4: 33240 queries in 60 seconds
5: 32460 queries in 60 seconds
avg: 32726.8
```


#### with queue:
benchmark-add ran 60s 5 times:
```
1: 30718 queries in 60 seconds
2: 32113 queries in 60 seconds
3: 32138 queries in 60 seconds
4: 32232 queries in 60 seconds
5: 32136 queries in 60 seconds
avg: 31867.4
```

avg performance difference with queue: -2.7%