@haadcode 2020-03-31
Thanks @tabcat, great to see the numbers! I'd be ok with that performance overhead.