@tabcat 2020-04-02
@haadcode do you think something like this *should* be implemented here or left for user space?
the third option would be making it optional which i like less than both former options but im also not opposed.