@aphelionz 2020-04-03
@tabcat sorry, maybe I'm being slow... What are the three options?