@tabcat 2020-04-03
sorry
1. don't implement queue
2. implement queue
3. make queue optional per store