@aphelionz 2020-04-03
I think @haadcode's comment was an approval, though not explicit.

I'm ok with merging this in as well