@aphelionz 2020-04-03
@tabcat Merged, congrats. Can you run the tests on `orbit-db` proper to make sure everything works?