@tabcat 2020-04-03
yes they are all passing however i wish i would have thought to tested that before now, my mistake. i only ran the orbit-db-store tests prior to this.
im sure you ran them both as well before merging, unfortunately i was afk at the time.